refactor(utils): Move common functions to utils/store
