+++ apps/common/models.py
@@ -1,27 +1,12 @@
 # pyright: reportUninitializedInstanceVariable=false
 import typing
 
-from django.core.exceptions import ValidationError
 from django.db import models
-from django.utils.translation import gettext
 from django_stubs_ext.db.models import TypedModelMeta
-from ulid import ULID
 
 from apps.user.models import User
 from main.db import Model
-
-
-def validate_ulid(val: str):
-    if val == "":
-        raise ValidationError(
-            gettext("Empty string is not a valid ULID value"),
-        )
-    try:
-        ULID.from_str(val)
-    except (ValueError, TypeError) as e:
-        raise ValidationError(
-            gettext("Not a valid ULID value '%s'") % (val),
-        ) from e
+from utils.common import validate_ulid
 
 
 # -- Abstracts
